requirements.txt
requests==2.31.0
python-dotenv==1.0.0
tomli==2.0.1
pytest==7.4.3
black==23.11.0 